% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptrue2.R
\name{prop_true}
\alias{prop_true}
\title{Find the proportion of x=1}
\usage{
prop_true(x)
}
\arguments{
\item{x}{}
}
\value{
Function will return the proportion of x = 1, in decimal form.
}
\description{
The prop_true function which finds the proportion of the value x=1 over a list of variables
# inputs:
}
\examples{
prop_true(x= c(FALSE,TRUE,TRUE)) #TRUE/FALSE data
prop_true(x=c(1,0,1,0)) #binary numerical data
prop_true(x= c(1,2,3)) #non-binary numerical data
}
